clearBuffers now can clear more than INT_MAX bytes

+++ platforms/cuda/include/CudaContext.h
@@ -311,7 +311,7 @@ public:
      * @param memory     the memory to clear
      * @param size       the size of the buffer in bytes
      */
-    void clearBuffer(CUdeviceptr memory, int size);
+    void clearBuffer(CUdeviceptr memory, size_t size);
     /**
      * Register a buffer that should be automatically cleared (all elements set to 0) at the start of each force or energy computation.
      */
@@ -322,7 +322,7 @@ public:
      * @param memory     the memory to clear
      * @param size       the size of the buffer in bytes
      */
-    void addAutoclearBuffer(CUdeviceptr memory, int size);
+    void addAutoclearBuffer(CUdeviceptr memory, size_t size);
     /**
      * Clear all buffers that have been registered with addAutoclearBuffer().
      */
@@ -590,7 +590,7 @@ private:
     std::vector<std::string> energyParamDerivNames;
     std::map<std::string, double> energyParamDerivWorkspace;
     std::vector<CUdeviceptr> autoclearBuffers;
-    std::vector<int> autoclearBufferSizes;
+    std::vector<size_t> autoclearBufferSizes;
     CudaIntegrationUtilities* integration;
     CudaExpressionUtilities* expression;
     CudaBondedUtilities* bonded;
