+++ platforms/cuda/src/CudaContext.cpp
@@ -756,8 +756,8 @@ void CudaContext::clearBuffer(ArrayInterface& array) {
     clearBuffer(unwrap(array).getDevicePointer(), array.getSize()*array.getElementSize());
 }
 
-void CudaContext::clearBuffer(CUdeviceptr memory, int size) {
-    int words = size/4;
+void CudaContext::clearBuffer(CUdeviceptr memory, size_t size) {
+    size_t words = size/4;
     void* args[] = {&memory, &words};
     executeKernel(clearBufferKernel, args, words, 128);
 }
@@ -766,7 +766,7 @@ void CudaContext::addAutoclearBuffer(ArrayInterface& array) {
     addAutoclearBuffer(unwrap(array).getDevicePointer(), array.getSize()*array.getElementSize());
 }
 
-void CudaContext::addAutoclearBuffer(CUdeviceptr memory, int size) {
+void CudaContext::addAutoclearBuffer(CUdeviceptr memory, size_t size) {
     autoclearBuffers.push_back(memory);
     autoclearBufferSizes.push_back(size/4);
 }
